+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed an issue where enabling log export configs did not account for lock errors.
+- Exporting Cloudwatch metrics on a serverless Cluster no longer generates an error.
+- Exporting Cloudwatch metrics while on the Basic plan now generates an error.
 
 ## [1.13.1]